requirements.txt
sentence-transformers>=2.2.2
faiss-cpu
numpy
tqdm
torch
